:art: rename and improve file structure
